requirements.txt
emailfinder
python-whois